Started clientserver branch (see Issue #4):
 * Created a new implementation of the protocol structures in ntlm2.py
 ** This uses ctypes structures for the message handling
 ** Supports creating and parsing all basic message structures
 ** the actual hashing functions still need to be done here
 * Created a new test_ntlm.py file that contains many tests for ntlm and ntlm2
 ** Moved tests out of ntlm __main__ section

The aim of this branch is to allow client- or server- code to be written
using the same classes. It should be possible to merge back to the
struct-packing code at some point, but this is helping me to organise
my thoughts.

I'm also planning to do ntlm2 support.

(I have the detailed commits in a local git repository, but didn't push
all of them - will make more fine-grained commits from here on)